# tsapi/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/device/": {
      "get": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to fetch all devices or devices by Hive ID or Device ID",
        "parameters": [
          {
            "name": "hiveId",
            "description": "Hive ID",
            "required": false,
            "in": "query",
            "type": "string"
          },
          {
            "name": "deviceId",
            "description": "Device ID like ( 13,21,33 or 22 )",
            "required": false,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Devices fetched successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "No devices found"
          },
          "500": {
            "description": "Failed to fetch devices"
          }
        }
      },
      "post": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to create a new device",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Device 1"
                },
                "hiveId": {
                  "type": "number",
                  "example": 1
                },
                "typeId": {
                  "type": "number",
                  "example": 3
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device already exists"
          },
          "201": {
            "description": "Device created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Failed to create device"
          }
        }
      },
      "put": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to update a device",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "deviceId": {
                  "type": "number",
                  "example": 7
                },
                "name": {
                  "type": "string",
                  "example": "Device 2"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device updated successfully"
          },
          "400": {
            "description": "No fields to update"
          },
          "404": {
            "description": "Updateble Device not found"
          },
          "500": {
            "description": "Failed to update device"
          }
        }
      },
      "delete": {
        "tags": [
          "Device"
        ],
        "description": "Endpoint to delete a device",
        "parameters": [
          {
            "name": "deviceId",
            "description": "Device ID",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Device deleted successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Deletable device not found"
          },
          "500": {
            "description": "Failed to delete device"
          }
        }
      }
    },
    "/api/hive/": {
      "get": {
        "tags": [
          "Hive"
        ],
        "description": "Endpoint to fetch all hives or hives by Hive ID or Area ID",
        "parameters": [
          {
            "name": "areaId",
            "description": "Area ID",
            "required": false,
            "in": "query",
            "type": "string"
          },
          {
            "name": "hiveId",
            "description": "Hive ID like ( 13,21,33 or 22 )",
            "required": false,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Hives fetched successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "No hives found"
          },
          "500": {
            "description": "Failed to fetch hives"
          }
        }
      },
      "post": {
        "tags": [
          "Hive"
        ],
        "description": "Endpoint to create a new hive",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Hive 10"
                },
                "areaId": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "name",
                "areaId"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Hive created successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "409": {
            "description": "Hive already exists"
          },
          "500": {
            "description": "Failed to create hive"
          }
        }
      },
      "put": {
        "tags": [
          "Hive"
        ],
        "description": "Endpoint to update a hive",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "hiveId": {
                  "type": "number",
                  "example": 7
                },
                "name": {
                  "type": "string",
                  "example": "Hive 20"
                },
                "areaId": {
                  "type": "number",
                  "example": 2
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Hive updated successfully"
          },
          "400": {
            "description": "No fields to update"
          },
          "404": {
            "description": "Updateble Hive not found"
          },
          "500": {
            "description": "Failed to update hive"
          }
        }
      },
      "delete": {
        "tags": [
          "Hive"
        ],
        "description": "Endpoint to delete a hive",
        "parameters": [
          {
            "name": "hiveId",
            "description": "Hive ID",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Hive deleted successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Deletable hive not found"
          },
          "500": {
            "description": "Failed to delete hive"
          }
        }
      }
    },
    "/api/area/": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "Endpoint to fetch all areas or areas by Area ID",
        "parameters": [
          {
            "name": "areaId",
            "description": "Area ID like ( 13,21,33 or 22 )",
            "required": false,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Areas fetched successfully"
          },
          "404": {
            "description": "No areas found"
          },
          "500": {
            "description": "Failed to fetch areas"
          }
        }
      },
      "post": {
        "tags": [
          "Area"
        ],
        "description": "Endpoint to add a new area",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "테스트"
                },
                "location": {
                  "type": "string",
                  "example": "37.375, 126.633"
                }
              },
              "required": [
                "name",
                "location"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Area added successfully"
          },
          "400": {
            "description": "Bad Request: Missing required fields"
          },
          "409": {
            "description": "Area already exists"
          },
          "500": {
            "description": "Failed to add area"
          }
        }
      },
      "put": {
        "tags": [
          "Area"
        ],
        "description": "Endpoint to update an area",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "areaId": {
                  "type": "number",
                  "example": 4
                },
                "name": {
                  "type": "string",
                  "example": "테스트2"
                },
                "location": {
                  "type": "string",
                  "example": "99.375, 999.633"
                }
              },
              "required": [
                "areaId",
                "name",
                "location"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Area updated successfully"
          },
          "400": {
            "description": "Bad Request: Missing required fields"
          },
          "404": {
            "description": "Area not found"
          },
          "500": {
            "description": "Failed to update area"
          }
        }
      },
      "delete": {
        "tags": [
          "Area"
        ],
        "description": "Endpoint to delete an area",
        "parameters": [
          {
            "name": "areaId",
            "description": "Area ID",
            "required": true,
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Area deleted successfully"
          },
          "400": {
            "description": "Bad Request: Missing required fields"
          },
          "404": {
            "description": "Area not found"
          },
          "500": {
            "description": "Failed to delete area"
          }
        }
      }
    }
  }
}